requirements.txt
aiogram
python-dotenv
aiohttp
pydantic
sqlalchemy[asyncio]
psutil